Fixed quick sort - ex1 - Quick sort

+++ src/main/ex1/quick_sort.h
@@ -11,7 +11,7 @@
  * @param compar It is the criterion according to which to sort the data (given two pointers to elements of the array, it returns a number greater than, equal to or less than zero if the first argument is respectively greater than, equal to or less than the second)
  * @return Index of the pivot
  */
-int partition(void *base, size_t size, size_t left, size_t right, int (*compar)(const void*, const void*));
+size_t partition(void *base, size_t size, size_t left, size_t right, int (*compar)(const void*, const void*));
 
 /**
  * @brief Quick sort of generic array
